Removed Grove from TO's dependency tree
