+++ lib/go-rfc/http.go
@@ -22,6 +22,7 @@ package rfc
 import (
 	"net/http"
 	"strings"
+	"time"
 
 	"github.com/apache/trafficcontrol/lib/go-util"
 )
@@ -53,4 +54,32 @@ func AcceptsGzip(r *http.Request) bool {
 		}
 	}
 	return false
+}
+
+// GetHTTPDate is a helper function which gets an HTTP date from the given map
+// (which is typically a `http.Header` or `CacheControl`.
+//
+// Returns false if the given key doesn't exist in the map, or if the value isn't
+// a valid HTTP Date per RFC2616§3.3.
+func GetHTTPDate(headers http.Header, key string) (time.Time, bool) {
+	maybeDate := headers.Get(key)
+	if maybeDate == "" {
+		return time.Time{}, false
+	}
+	return ParseHTTPDate(maybeDate)
+}
+
+// ParseHTTPDate parses the given RFC7231§7.1.1 HTTP-date.
+func ParseHTTPDate(d string) (time.Time, bool) {
+	if t, err := time.Parse(time.RFC1123, d); err == nil {
+		return t, true
+	}
+	if t, err := time.Parse(time.RFC850, d); err == nil {
+		return t, true
+	}
+	if t, err := time.Parse(time.ANSIC, d); err == nil {
+		return t, true
+	}
+	return time.Time{}, false
+
 }
