+++ traffic_ops/traffic_ops_golang/api/generic_crud.go
@@ -23,7 +23,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"github.com/apache/trafficcontrol/grove/web"
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-rfc"
 	ims "github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/util/ims"
@@ -142,7 +141,7 @@ func TryIfModifiedSinceQuery(val GenericReader, h http.Header, where string, ord
 	if len(imsDateHeader) == 0 {
 		return runSecond, max
 	}
-	if imsDate, ok = web.ParseHTTPDate(imsDateHeader[0]); !ok {
+	if imsDate, ok = rfc.ParseHTTPDate(imsDateHeader[0]); !ok {
 		log.Warnf("IMS request header date '%s' not parsable", imsDateHeader[0])
 		return runSecond, max
 	}
